src/challenge_2/requirements.txt
flask
pycrypto
gunicorn
